+++ plantseg/core/__init__.py
@@ -1,4 +1,3 @@
 from plantseg.core.image import ImageProperties, PlantSegImage
-from plantseg.core.voxelsize import VoxelSize
 
-__all__ = ['PlantSegImage', 'VoxelSize', 'ImageProperties']
+__all__ = ['PlantSegImage', 'ImageProperties']
